RavenDB-21074 - Get the new id when updating huge documents
