+++ src/Raven.Server/Documents/Handlers/BatchHandler.cs
@@ -959,7 +959,7 @@ namespace Raven.Server.Documents.Handlers
                                 throw;
                             }
 
-                            context.DocumentDatabase.HugeDocuments.AddIfDocIsHuge(cmd.Id, cmd.Document.Size);
+                            context.DocumentDatabase.HugeDocuments.AddIfDocIsHuge(putResult.Id, cmd.Document.Size);
                             AddPutResult(putResult);
                             lastPutResult = putResult;
                             break;
